feat: write board form page funnel 적용
