--- src/app/create/components/AgeModal/getAgeList.ts
@@ -1,10 +1,0 @@
-const getAgeList = (min: number, max: number) => {
-  const list = [''];
-  for (let i = min; i < max; i++) {
-    list.push(`${i.toString()}세`);
-  }
-
-  return list;
-};
-
-export default getAgeList;
