--- src/app/create/style.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-export const formWrapper = style({
-  width: '100%',
-  display: 'grid',
-  gridAutoFlow: 'row',
-  gap: '3.5rem',
-});
