--- src/app/create/utils/getTimes.ts
@@ -1,18 +1,0 @@
-function getTimeList() {
-  const timeList = [];
-
-  for (let hour = 12; hour <= 23; hour++) {
-    for (let minute = 0; minute <= 45; minute += 15) {
-      const time = '오후 ' + padTime(hour) + ':' + padTime(minute);
-      timeList.push(time);
-    }
-  }
-
-  return timeList;
-}
-
-function padTime(value: number) {
-  return value.toString().padStart(2, '0');
-}
-
-export default getTimeList;
